requirements.txt
# Core
streamlit==1.28.0
pandas==2.1.4
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Visualization
plotly==5.17.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3